requirements.txt
pandas
numpy
matplotlib
Scikit-learn
tensorflow==2.16.1
streamlit
scikeras
Pillow
opencv-python-headless
plotly